Move ExceptionsOperation to separate file
